move the contract as well
